healthai-API/MSPR1-ETL/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic==2.7.1
email-validator==2.1.1
python-dotenv==1.0.1
gunicorn==23.0.0